:violin: Refactor functions for better readability
