+++ lib/onlineServices.js
@@ -1,86 +1,56 @@
-function getBookingSystem(gp, data) {
-  const bookingSystem = { supplier: data.Supplier };
-  let bookOnlineLink;
+const systems = require('./systems');
 
-  switch (bookingSystem.supplier) {
-    case 'EMIS':
-      bookOnlineLink = 'https://patient.emisaccess.co.uk/appointments/available';
-      break;
-    case 'Informatica':
-    case 'INPS':
-      bookOnlineLink = 'https://www.patient-services.co.uk/web/ps/login';
-      break;
-    case 'Microtest':
-      bookOnlineLink = 'https://www.thewaiting-room.net/';
-      break;
-    case 'TPP':
-      bookOnlineLink = `https://systmonline.tpp-uk.com/Login?PracticeId=${gp.odsCode}`;
-      break;
-    case 'NK':
-    default:
-      bookOnlineLink = gp.contact.website;
-      break;
-  }
-  bookingSystem.bookOnlineLink = bookOnlineLink;
-  return bookingSystem;
+function findByOdsCode(data, odsCode) {
+  return data.find(item => item.GPPracticeCode === odsCode);
 }
 
-function getRepeatPrescriptionSystem(gp, data) {
-  const system = { supplier: data.Supplier };
-  let url;
+function getBookingSystem(gp, match) {
+  return systems.getBookingSystem(gp, match);
+}
 
-  switch (system.supplier) {
-    case 'EMIS':
-      url = 'https://patient.emisaccess.co.uk/prescriptions/request';
+function getCodedRecordsSystem(gp, match) {
+  return systems.getCodedRecordsSystem(gp, match);
+}
+
+function getRepeatPrescriptionSystem(gp, match) {
+  return systems.getRepeatPrescriptionSystem(gp, match);
+}
+
+function getSupplierSystem(config, match) {
+  const gp = config.gp;
+  let system;
+
+  switch (config.key) {
+    case 'appointments':
+      system = getBookingSystem(gp, match);
       break;
-    case 'Informatica':
-    case 'INPS':
-      url = 'https://www.patient-services.co.uk/web/ps/login';
+    case 'repeatPrescriptions':
+      system = getRepeatPrescriptionSystem(gp, match);
       break;
-    case 'Microtest':
-      url = 'https://www.thewaiting-room.net/';
+    case 'codedRecords':
+      system = getCodedRecordsSystem(gp, match);
       break;
-    case 'TPP':
-      url = `https://systmonline.tpp-uk.com/Login?PracticeId=${gp.odsCode}`;
-      break;
-    case 'NK':
     default:
-      url = gp.contact.website;
-      break;
+      throw new Error(`Unknown key: ${config.key}`);
   }
-  system.url = url;
   return system;
 }
 
-function getCodedRecordsSystem(gp, data) {
-  const system = { supplier: data.Supplier };
-  let url;
+function add(config) {
+  const gp = config.gp;
+  const match = findByOdsCode(config.systemList, gp.odsCode);
 
-  switch (system.supplier) {
-    case 'EMIS':
-      url = 'https://patient.emisaccess.co.uk/appointments/available';
-      break;
-    case 'Informatica':
-    case 'INPS':
-      url = 'https://www.patient-services.co.uk/web/ps/login';
-      break;
-    case 'Microtest':
-      url = 'https://www.thewaiting-room.net/';
-      break;
-    case 'TPP':
-      url = `https://systmonline.tpp-uk.com/Login?PracticeId=${gp.odsCode}`;
-      break;
-    case 'NK':
-    default:
-      url = gp.contact.website;
-      break;
+  if (match) {
+    const system = getSupplierSystem(config, match);
+    if (config.key === 'appointments') {
+    // eslint-disable-next-line no-param-reassign
+      gp.bookingSystem = system;
+    }
+    // eslint-disable-next-line no-param-reassign
+    gp.onlineServices[config.key] = system;
   }
-  system.url = url;
-  return system;
 }
 
 module.exports = {
-  getBookingSystem,
-  getCodedRecordsSystem,
-  getRepeatPrescriptionSystem,
+  add,
 };
